components folder created. Agent moved here, moving the other components now
