+++ src/aurite/workflows/simple_workflow.py
@@ -18,7 +18,7 @@ from ..config.config_models import (
 # Import AgentExecutionResult for type hinting the result from facade.run_agent()
 if TYPE_CHECKING:
     from ..execution.facade import ExecutionFacade
-    from ..agents.agent_models import AgentExecutionResult  # Added import
+    from ..components.components.agents.agent_models import AgentExecutionResult  # Added import
 
 logger = logging.getLogger(__name__)
